Add folder parameter and README (#5)
There was a question about the place the files are saved. The files are by
default saved in the folder, where the script is running - but I would guess
this is not really easily apparent to the non-programmers. So I added the
parameter, where the user can define in which folder are the files.